fix of button color: it was changed because of last change
This change occurred because the XML Theme was removed. Without this
theme, the application no longer recognizes 'Dark Mode,' causing the
buttons to revert to their default 'Light Mode' text color, which is a
deep purple. As the background is manually set to black, this deep
purple text is currently not visible. To rectify this, we have manually
adjusted the text color on your buttons to either white or green.